Static content describing known limitations of rbac
