--- cypher/cypher-docs/src/docs/dev/ql/administration/security/security-and-indexes-schema.asciidoc
@@ -1,1 +1,0 @@
-For single property and composite indexes, queries that
